extract navigation buttons to a module
