+++ modules/navigation-buttons/edit-home.js
@@ -1,21 +1,17 @@
-/**
- * @flow
- * Exports a button that opens the Edit Home screen
- */
+// @flow
 
 import * as React from 'react'
 import {Text} from 'react-native'
 import {Touchable} from '@frogpond/touchable'
-import type {NavType} from '../../views/types'
+import type {NavigationScreenProp} from 'react-navigation'
 import {commonStyles} from './styles'
 
-export function EditHomeButton({
-	navigation,
-	buttonStyle,
-}: {
-	navigation: NavType,
+type Props = {
+	navigation: NavigationScreenProp<*>,
 	buttonStyle?: any,
-}) {
+}
+
+export function EditHomeButton({navigation, buttonStyle}: Props) {
 	return (
 		<Touchable
 			accessibilityComponentType="button"
