+++ modules/navigation-buttons/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@frogpond/navigation-buttons",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "author": "",
+  "license": "ISC",
+  "scripts": {
+    "test": "jest"
+  },
+  "peerDependencies": {
+    "@frogpond/app-theme": "^1.0.0",
+    "react": "^16.0.0",
+    "react-native": "^0.55.4",
+    "react-native-vector-icons": "^4.0.0",
+    "react-navigation": "^2.11.2"
+  }
+}
